// src/app/features/transfer-channel/citad/hold-receiver-name/modal/interface.ts
export interface IResponseHoldReceiverName {
    id: number;
    key: string;
    value: {
        $type: string;
        holdReceiverNamePatterns: string[]
    };
    active: boolean;
    reason: string;
    createdBy: string;
    createdAt: string;
    updatedBy: string;
    updatedAt: string;
}

export interface PaginationHistoryHoldReceiverNameDto<T> {
    content: T[];
    pageable: {
        pageNumber: number;
        pageSize: number;
        sort: {
            empty: boolean;
            sorted: boolean;
            unsorted: boolean;
        };
        offset: number;
        unpaged: boolean;
        paged: boolean;
    };
    totalPages?: number;
    totalElements: number;
    last: boolean;
    size: number;
    number: number;
    sort: {
        empty: boolean;
        sorted: boolean;
        unsorted: boolean;
    };
    numberOfElements: number;
    first: boolean;
    empty: boolean;
}

export interface IResponseHistoryHoldReceiverName {
    id: number;
    configId: number;
    key: string;
    active: boolean;
    reason: string;
    action: string;
    createdBy: string;
    value: {
        $type: string;
        abbreviations: {
            "DOANH NGHIEP TN": string[];
            "Doanh nghiệp Trang": string[];
            "DNTN": []
        }
    };
    createdAt: string;
    updatedBy: string;
    updatedAt: string;
}

export interface IRequestUpdateHoldReceiverName {
    key: string,
    value: {
        $type: string,
        holdReceiverNamePatterns: string[]
    },
    reason: string
}